Import of CDENT/Purple-1.0 from CPAN.
gitpan-cpan-distribution: Purple
gitpan-cpan-version:      1.0
gitpan-cpan-path:         CDENT/Purple/Purple-1.0.tar.gz
gitpan-cpan-author:       CDENT
gitpan-cpan-maturity:     released

+++ t/client.t
@@ -65,5 +65,5 @@ sub get_server {
 }
 
 END {
-    kill 15, $pid;
+    kill 15, $pid if defined $pid;
 }
